% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenMonteCarloSamples.R
\name{GenMonteCarloSamples}
\alias{GenMonteCarloSamples}
\title{GenMonteCarloSamples}
\usage{
GenMonteCarloSamples(calc, calc_args, iterations = 100)
}
\arguments{
\item{calc}{A function to sample}

\item{calc_args}{A function to create arguments for the func being sampled}

\item{iterations}{The number of samples to create}
}
\value{
samples The return values from the calc function
}
\description{
\code{GenMonteCarloSamples} returns a list of samples from executing the function
with arguments
}
\examples{

calc_estimate <- function(a, ef) {
  return(a * ef)
}
calc_args <- function() {
  area <- 10
  emission_factor <- rtriangle(
    n = 1, theta = ef_defor,
    lower = ef_defor_lci,
    upper = ef_defor_uci
  )
  return(list(area, emission_factor))
}
GenMonteCarloSamples(calc_estimate, calc_args, 1000) # returns object
}
