% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMonteCarloSamples.R
\name{PlotMonteCarloSamples}
\alias{PlotMonteCarloSamples}
\title{PlotMonteCarloSamples}
\usage{
PlotMonteCarloSamples(o, est, title = "")
}
\arguments{
\item{o}{Debug output from StableMonteCarloSamples}
}
\value{
graph of the stable analysis
}
\description{
\code{PlotMonteCarloSamples} plots the stable analysis
}
